move base to simile-ajax-base
